# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sistema de facturación para lavandería
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Cliente {
  id: number;
  nombre: string;
  /** @nullable */
  apellido?: string | null;
  telefono: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  direccion?: string | null;
  /** @nullable */
  notas?: string | null;
  createdAt: string;
}

export interface ClienteInput {
  /** @minLength 1 */
  nombre: string;
  apellido?: string;
  /** @minLength 1 */
  telefono: string;
  email?: string;
  direccion?: string;
  notas?: string;
}

export interface ClienteUpdate {
  /** @minLength 1 */
  nombre?: string;
  apellido?: string;
  telefono?: string;
  email?: string;
  direccion?: string;
  notas?: string;
}

export interface Servicio {
  id: number;
  nombre: string;
  /** @nullable */
  descripcion?: string | null;
  precio: number;
  /** Ej: por kilo, por prenda, por par */
  unidad: string;
  activo: boolean;
}

export interface ServicioInput {
  /** @minLength 1 */
  nombre: string;
  descripcion?: string;
  /** @minimum 0 */
  precio: number;
  /** @minLength 1 */
  unidad: string;
  activo?: boolean;
}

export interface ServicioUpdate {
  /** @minLength 1 */
  nombre?: string;
  descripcion?: string;
  /** @minimum 0 */
  precio?: number;
  unidad?: string;
  activo?: boolean;
}

export type OrdenEstado = typeof OrdenEstado[keyof typeof OrdenEstado];


export const OrdenEstado = {
  pendiente: 'pendiente',
  en_proceso: 'en_proceso',
  listo: 'listo',
  entregado: 'entregado',
} as const;

export interface Orden {
  id: number;
  clienteId: number;
  /** @nullable */
  clienteNombre?: string | null;
  estado: OrdenEstado;
  total: number;
  /** @nullable */
  notas?: string | null;
  /** @nullable */
  fechaEntregaEstimada?: string | null;
  /** @nullable */
  fechaEntregaReal?: string | null;
  createdAt: string;
}

export type OrdenConDetalleEstado = typeof OrdenConDetalleEstado[keyof typeof OrdenConDetalleEstado];


export const OrdenConDetalleEstado = {
  pendiente: 'pendiente',
  en_proceso: 'en_proceso',
  listo: 'listo',
  entregado: 'entregado',
} as const;

export interface OrdenItem {
  id: number;
  ordenId: number;
  servicioId: number;
  /** @nullable */
  servicioNombre?: string | null;
  cantidad: number;
  precioUnitario: number;
  subtotal: number;
}

export type FacturaEstadoPago = typeof FacturaEstadoPago[keyof typeof FacturaEstadoPago];


export const FacturaEstadoPago = {
  pendiente: 'pendiente',
  pagado: 'pagado',
} as const;

/**
 * @nullable
 */
export type FacturaMetodoPago = typeof FacturaMetodoPago[keyof typeof FacturaMetodoPago] | null;


export const FacturaMetodoPago = {
  efectivo: 'efectivo',
  tarjeta: 'tarjeta',
  transferencia: 'transferencia',
} as const;

export interface Factura {
  id: number;
  ordenId: number;
  numeroFactura: string;
  total: number;
  estadoPago: FacturaEstadoPago;
  /** @nullable */
  metodoPago?: FacturaMetodoPago;
  /** @nullable */
  fechaPago?: string | null;
  createdAt: string;
}

export interface OrdenConDetalle {
  id: number;
  clienteId: number;
  /** @nullable */
  clienteNombre?: string | null;
  /** @nullable */
  clienteTelefono?: string | null;
  estado: OrdenConDetalleEstado;
  total: number;
  /** @nullable */
  notas?: string | null;
  /** @nullable */
  fechaEntregaEstimada?: string | null;
  /** @nullable */
  fechaEntregaReal?: string | null;
  createdAt: string;
  items: OrdenItem[];
  factura?: Factura;
}

export interface OrdenItemInput {
  servicioId: number;
  /** @minimum 0.01 */
  cantidad: number;
  /** Opcional; si se omite se usa el precio del servicio */
  precioUnitario?: number;
}

export interface OrdenInput {
  clienteId: number;
  notas?: string;
  fechaEntregaEstimada?: string;
  items?: OrdenItemInput[];
}

export type OrdenUpdateEstado = typeof OrdenUpdateEstado[keyof typeof OrdenUpdateEstado];


export const OrdenUpdateEstado = {
  pendiente: 'pendiente',
  en_proceso: 'en_proceso',
  listo: 'listo',
  entregado: 'entregado',
} as const;

export interface OrdenUpdate {
  estado?: OrdenUpdateEstado;
  notas?: string;
  fechaEntregaEstimada?: string;
  fechaEntregaReal?: string;
}

export type FacturaConDetalleEstadoPago = typeof FacturaConDetalleEstadoPago[keyof typeof FacturaConDetalleEstadoPago];


export const FacturaConDetalleEstadoPago = {
  pendiente: 'pendiente',
  pagado: 'pagado',
} as const;

/**
 * @nullable
 */
export type FacturaConDetalleMetodoPago = typeof FacturaConDetalleMetodoPago[keyof typeof FacturaConDetalleMetodoPago] | null;


export const FacturaConDetalleMetodoPago = {
  efectivo: 'efectivo',
  tarjeta: 'tarjeta',
  transferencia: 'transferencia',
} as const;

export interface FacturaConDetalle {
  id: number;
  ordenId: number;
  numeroFactura: string;
  total: number;
  estadoPago: FacturaConDetalleEstadoPago;
  /** @nullable */
  metodoPago?: FacturaConDetalleMetodoPago;
  /** @nullable */
  fechaPago?: string | null;
  createdAt: string;
  orden: OrdenConDetalle;
}

export interface FacturaInput {
  ordenId: number;
}

export type PagoInputMetodoPago = typeof PagoInputMetodoPago[keyof typeof PagoInputMetodoPago];


export const PagoInputMetodoPago = {
  efectivo: 'efectivo',
  tarjeta: 'tarjeta',
  transferencia: 'transferencia',
} as const;

export interface PagoInput {
  metodoPago: PagoInputMetodoPago;
}

export interface DashboardResumen {
  ordenesHoy: number;
  ordenesPendientes: number;
  ordenesEnProceso: number;
  ordenesListas: number;
  ingresosHoy: number;
  ingresosMes: number;
  clientesTotales: number;
  facturasPendientes: number;
  facturasPendientesMonto?: number;
}

export type ListClientesParams = {
search?: string;
};

export type ListOrdenesParams = {
estado?: ListOrdenesEstado;
clienteId?: number;
};

export type ListOrdenesEstado = typeof ListOrdenesEstado[keyof typeof ListOrdenesEstado];


export const ListOrdenesEstado = {
  pendiente: 'pendiente',
  en_proceso: 'en_proceso',
  listo: 'listo',
  entregado: 'entregado',
} as const;

export type ListFacturasParams = {
estadoPago?: ListFacturasEstadoPago;
};

export type ListFacturasEstadoPago = typeof ListFacturasEstadoPago[keyof typeof ListFacturasEstadoPago];


export const ListFacturasEstadoPago = {
  pendiente: 'pendiente',
  pagado: 'pagado',
} as const;