# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sistema de facturación para lavandería
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clientes:
    get:
      operationId: listClientes
      tags: [clientes]
      summary: Listar clientes
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
    post:
      operationId: createCliente
      tags: [clientes]
      summary: Crear cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClienteInput"
      responses:
        "201":
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
  /clientes/{id}:
    get:
      operationId: getCliente
      tags: [clientes]
      summary: Obtener cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCliente
      tags: [clientes]
      summary: Actualizar cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClienteUpdate"
      responses:
        "200":
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
    delete:
      operationId: deleteCliente
      tags: [clientes]
      summary: Eliminar cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Eliminado
  /clientes/{id}/historial:
    get:
      operationId: getClienteHistorial
      tags: [clientes]
      summary: Historial de órdenes del cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Historial de órdenes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Orden"

  # ─── Servicios ───────────────────────────────────────────────────────────────
...
  /facturas:
    get:
      operationId: listFacturas
      tags: [facturas]
      summary: Listar facturas
      parameters:
        - name: estadoPago
          in: query
          schema:
            type: string
            enum: [pendiente, pagado]
      responses:
        "200":
          description: Lista de facturas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Factura"
    post:
      operationId: createFactura
      tags: [facturas]
      summary: Crear factura para una orden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacturaInput"
      responses:
        "201":
          description: Factura creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Factura"
  /facturas/{id}:
    get:
      operationId: getFactura
      tags: [facturas]
      summary: Obtener factura
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Factura
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacturaConDetalle"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /facturas/{id}/pagar:
    patch:
      operationId: pagarFactura
      tags: [facturas]
      summary: Marcar factura como pagada
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PagoInput"
      responses:
        "200":
          description: Factura pagada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Factura"

  # ─── Dashboard ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Cliente ───────────────────────────────────────────────────────────────
    Cliente:
      type: object
      required: [id, nombre, telefono, createdAt]
      properties:
        id:
          type: number
        nombre:
          type: string
        apellido:
          type: ["string", "null"]
        telefono:
          type: string
        email:
          type: ["string", "null"]
        direccion:
          type: ["string", "null"]
        notas:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ClienteInput:
      type: object
      required: [nombre, telefono]
      properties:
        nombre:
          type: string
          minLength: 1
        apellido:
          type: string
        telefono:
          type: string
          minLength: 1
        email:
          type: string
        direccion:
          type: string
        notas:
          type: string
    ClienteUpdate:
      type: object
      properties:
        nombre:
          type: string
          minLength: 1
        apellido:
          type: string
        telefono:
          type: string
        email:
          type: string
        direccion:
          type: string
        notas:
          type: string

    # ─── Servicio ──────────────────────────────────────────────────────────────
...
    Orden:
      type: object
      required: [id, clienteId, estado, total, createdAt]
      properties:
        id:
          type: number
        clienteId:
          type: number
        clienteNombre:
          type: ["string", "null"]
        estado:
          type: string
          enum: [pendiente, en_proceso, listo, entregado]
        total:
          type: number
        notas:
          type: ["string", "null"]
        fechaEntregaEstimada:
          type: ["string", "null"]
          format: date-time
        fechaEntregaReal:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    OrdenConDetalle:
      type: object
      required: [id, clienteId, estado, total, items, createdAt]
      properties:
        id:
          type: number
        clienteId:
          type: number
        clienteNombre:
          type: ["string", "null"]
        clienteTelefono:
          type: ["string", "null"]
        estado:
          type: string
          enum: [pendiente, en_proceso, listo, entregado]
        total:
          type: number
        notas:
          type: ["string", "null"]
        fechaEntregaEstimada:
          type: ["string", "null"]
          format: date-time
        fechaEntregaReal:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrdenItem"
        factura:
          $ref: "#/components/schemas/Factura"
...
    OrdenItem:
      type: object
      required: [id, ordenId, servicioId, cantidad, precioUnitario, subtotal]
      properties:
        id:
          type: number
        ordenId:
          type: number
        servicioId:
          type: number
        servicioNombre:
          type: ["string", "null"]
        cantidad:
          type: number
        precioUnitario:
          type: number
        subtotal:
          type: number
...
    Factura:
      type: object
      required: [id, ordenId, numeroFactura, total, estadoPago, createdAt]
      properties:
        id:
          type: number
        ordenId:
          type: number
        numeroFactura:
          type: string
        total:
          type: number
        estadoPago:
          type: string
          enum: [pendiente, pagado]
        metodoPago:
          type: ["string", "null"]
          enum: [efectivo, tarjeta, transferencia, null]
        fechaPago:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    FacturaConDetalle:
      type: object
      required: [id, ordenId, numeroFactura, total, estadoPago, createdAt, orden]
      properties:
        id:
          type: number
        ordenId:
          type: number
        numeroFactura:
          type: string
        total:
          type: number
        estadoPago:
          type: string
          enum: [pendiente, pagado]
        metodoPago:
          type: ["string", "null"]
          enum: [efectivo, tarjeta, transferencia, null]
        fechaPago:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        orden:
          $ref: "#/components/schemas/OrdenConDetalle"
    FacturaInput:
      type: object
      required: [ordenId]
      properties:
        ordenId:
          type: number
    PagoInput:
      type: object
      required: [metodoPago]
      properties:
        metodoPago:
          type: string
          enum: [efectivo, tarjeta, transferencia]

    # ─── Dashboard ─────────────────────────────────────────────────────────────
...