# lib/api-zod/src/generated/types/facturaConDetalle.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sistema de facturación para lavandería
 * OpenAPI spec version: 0.1.0
 */
import type { FacturaConDetalleEstadoPago } from './facturaConDetalleEstadoPago';
import type { FacturaConDetalleMetodoPago } from './facturaConDetalleMetodoPago';
import type { OrdenConDetalle } from './ordenConDetalle';

export interface FacturaConDetalle {
  id: number;
  ordenId: number;
  numeroFactura: string;
  total: number;
  estadoPago: FacturaConDetalleEstadoPago;
  /** @nullable */
  metodoPago?: FacturaConDetalleMetodoPago;
  /** @nullable */
  fechaPago?: Date | null;
  createdAt: Date;
  orden: OrdenConDetalle;
}