# lib/api-zod/src/generated/types/orden.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sistema de facturación para lavandería
 * OpenAPI spec version: 0.1.0
 */
import type { OrdenEstado } from './ordenEstado';

export interface Orden {
  id: number;
  clienteId: number;
  /** @nullable */
  clienteNombre?: string | null;
  estado: OrdenEstado;
  total: number;
  /** @nullable */
  notas?: string | null;
  /** @nullable */
  fechaEntregaEstimada?: Date | null;
  /** @nullable */
  fechaEntregaReal?: Date | null;
  createdAt: Date;
}