# lib/api-zod/src/generated/types/pagoInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sistema de facturación para lavandería
 * OpenAPI spec version: 0.1.0
 */
import type { PagoInputMetodoPago } from './pagoInputMetodoPago';

export interface PagoInput {
  metodoPago: PagoInputMetodoPago;
}